Golang: Replacing gocql dependency with nuid
github.com/nats-io/nuid is already required for github.com/nats-io/nats which is used by this repo.

Replacing this dependency removes the necessity of vendoring a cassandra client for unique identifier generation.

For reference, the NUID library description is quoted as follows:

> NUID needs to be very fast to generate and truly unique, all while being entropy pool friendly.
> We will use 12 bytes of crypto generated data (entropy draining), and 10 bytes of sequential data
> that is started at a pseudo random number and increments with a pseudo-random increment.
> Total is 22 bytes of base 62 ascii text :)

This seems to be a suitable replacement for UUIDs.